+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Button to download the perfs of multiple charts at once (#312)
 -   Buttons to download the tested models in a test task drawer (#319)
+-   Hierarchy background to PerfBrowser compute plan filters (#321)
 
 ### Fixed